购物车预购数量修改

+++ service/src/main/java/com/info/service/ShoppingService.java
@@ -11,5 +11,5 @@ public interface ShoppingService {
     List<GoodsEntity> selectOrderGoodsList(Map<String ,Object> map) throws Exception;
     void deleteOrderRecord(Long goodsId,Long userId) throws Exception;
     void payment(String type, Long goodsId,Long userId);
-    void saveOrder(Long userId, Long goodsId, Long orderNumber);
+    void saveOrder(Long userId, Long goodsId, int orderNumber);
 }
